# prompt/API_SPEC.yaml
openapi: 3.0.3
info:
  title: WireGuard Panel API
  description: |
    A RESTful API for managing a WireGuard server panel.
    The API has a three-level hierarchy: Interfaces -> Servers -> Clients.
    Authentication is handled via session cookies.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Development server (Assuming default Site URL Prefix '/' and API Prefix '/api')
...
paths:
  /service/login:
    post:
      tags:
        - Service
      summary: Log in to the service
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful. Session cookie is set.
          headers:
            Set-Cookie:
              description: Session token cookie.
              schema:
                type: string
        '401':
          description: Invalid credentials.
  /service/logout:
    post:
      tags:
        - Service
      summary: Log out from the service
      operationId: logout
      responses:
        '204':
          description: Logout successful. Session cookie is cleared.
      security:
        - cookieAuth: []
  /service/config:
    get:
      tags:
        - Service
      summary: Get Service Configuration
      description: Retrieves the current service configuration, excluding sensitive data like the password hash.
      operationId: getServiceConfig
      responses:
        '200':
          description: Successful retrieval of service configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceConfig'
      security:
        - cookieAuth: []
  /service/password:
    put:
      tags:
        - Service
      summary: Set a New Password
      description: Updates the login password for the service. The change is persisted to the configuration file.
      operationId: setServicePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  format: password
                  description: The new password.
              required:
                - password
      responses:
        '204':
          description: Password updated successfully.
        '401':
          description: Unauthorized.
      security:
        - cookieAuth: []
  /service/snatroamingoffsetvalid:
    get:
      tags:
        - Service
      summary: Check SNAT Roaming Offset Validity
      description: |
        Validates if a SNAT roaming offset is valid for a given master interface and address family.
        Returns the type of NAT mapping (SNAT or NETMAP) and the mapped network.
      operationId: checkSNATRoamingOffsetValid
      parameters:
        - name: af
          in: query
          required: true
          schema:
            type: string
            enum: ["4", "6"]
          description: Address family - "4" for IPv4, "6" for IPv6
        - name: ifname
          in: query
          required: true
          schema:
            type: string
          description: The master network interface name to use for roaming SNAT
        - name: offset
          in: query
          required: true
          schema:
            type: string
          description: The IP or network offset in CIDR notation
      responses:
        '200':
          description: Offset validation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    enum: ["SNAT", "NETMAP"]
                    description: The type of NAT mapping
                  mapped network:
                    type: string
                    description: The resulting mapped network in CIDR notation
                example:
                  type: "SNAT"
                  mapped network: "192.168.1.0/24"
        '400':
          description: Invalid parameters or validation failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                  error_params:
                    type: string
                    enum: ["ifname", "offset", "addressfamily"]
                    description: Which parameter caused the error
                example:
                  error: "Invalid offset: network mask too large"
                  error_params: "offset"
      security:
        - cookieAuth: []
  /interfaces:
    get:
      tags:
        - Interfaces
      summary: List all Interfaces
      operationId: listInterfaces
      responses:
        '200':
          description: An array of WireGuard interfaces.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Interface'
      security:
        - cookieAuth: []
    post:
      tags:
        - Interfaces
      summary: Create a new Interface
      operationId: createInterface
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterfaceCreate'
      responses:
        '201':
          description: Interface created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interface'
        '409':
          description: Conflict, e.g. ifname already exists.
      security:
        - cookieAuth: []
  /interfaces/{ifId}:
    parameters:
      - $ref: '#/components/parameters/ifId'
    get:
      tags:
        - Interfaces
      summary: Get Interface Details
      operationId: getInterface
      responses:
        '200':
          description: The details of the specified interface, including its public key and a full list of its associated servers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interface'
        '404':
          description: Interface not found.
      security:
        - cookieAuth: []
    put:
      tags:
        - Interfaces
      summary: Edit an Interface
      description: |
        Edits an interface's properties.
        - Changing `mtu` will apply `ip link set dev $IFNAME mtu $NEW_MTU`.
        - Changing `vrfName`, `fwMark`, `port`, or `privateKey` will trigger a regeneration and sync of the WireGuard config file (`wg syncconf`).
        - When `vrfName` is changed, a check is performed to ensure no child server's network overlaps with another network in the target VRF.
      operationId: editInterface
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterfaceUpdate'
      responses:
        '200':
          description: Interface updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interface'
        '404':
          description: Interface not found.
        '409':
          description: Conflict, e.g. VRF network overlap detected.
      security:
        - cookieAuth: []
    delete:
      tags:
        - Interfaces
      summary: Delete an Interface
      description: Deletes an interface and all of its associated servers and clients.
      operationId: deleteInterface
      responses:
        '204':
          description: Interface deleted successfully.
        '404':
          description: Interface not found.
      security:
        - cookieAuth: []
...
  /interfaces/{ifId}/servers:
    parameters:
      - $ref: '#/components/parameters/ifId'
    get:
      tags:
        - Servers
      summary: List Servers for an Interface
      operationId: listServers
      responses:
        '200':
          description: An array of servers for the specified interface.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
        '404':
          description: Interface not found.
      security:
        - cookieAuth: []
    post:
      tags:
        - Servers
      summary: Create a new Server
      operationId: createServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '201':
          description: Server created successfully. New servers are always created as disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: Invalid input (e.g., neither IPv4 nor IPv6 enabled).
        '404':
          description: Interface not found.
        '409':
          description: Conflict, e.g. network overlaps with another server in the same VRF.
      security:
        - cookieAuth: []
  /interfaces/{ifId}/servers/{serverId}:
    parameters:
      - $ref: '#/components/parameters/ifId'
      - $ref: '#/components/parameters/serverId'
    get:
      tags:
        - Servers
      summary: Get Server Details
      operationId: getServer
      responses:
        '200':
          description: The details of the specified server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          description: Server or Interface not found.
      security:
        - cookieAuth: []
    put:
      tags:
        - Servers
      summary: Edit a Server
      description: |
        Edits server properties.
        - If `network` is changed, a check is performed to ensure all existing client IPs are still valid within the new network range.
        - If `network`, `routedNetworks`, or `snat` settings are changed, old firewall rules are removed and new ones are applied.
      operationId: editServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
      responses:
        '200':
          description: Server updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '404':
          description: Server or Interface not found.
        '409':
          description: Conflict, e.g., client IP shifting results in an out-of-range address, or network overlap.
      security:
        - cookieAuth: []
    delete:
      tags:
        - Servers
      summary: Delete a Server
      description: Disables the server (removes IPs, firewall rules, clients from wg conf), then deletes the server and all its associated clients from the configuration.
      operationId: deleteServer
      responses:
        '204':
          description: Server deleted successfully.
        '404':
          description: Server or Interface not found.
      security:
        - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
  parameters:
    ifId:
      name: ifId
      in: path
      required: true
      schema:
        type: string
      description: The unique ID of the interface.
    serverId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
      description: The unique ID of the server.
...
  schemas:
...
    ServiceConfig:
      type: object
      properties:
        wireguardConfigPath:
          type: string
          description: The path to store WireGuard interface configs.
          example: "/etc/wireguard"
        user:
          type: string
          description: The login username.
          example: "admin"
        listenIP:
          type: string
          description: The listen IP for the service.
          example: "0.0.0.0"
        listenPort:
          type: integer
          description: The listen port for the service.
          example: 8080
        siteUrlPrefix:
          type: string
          description: The prefix for the site URL. Default is '/'.
          example: "/"
        apiPrefix:
          type: string
          description: The prefix for backend APIs, appended to siteUrlPrefix. Default is '/api'.
          example: "/api"
        panelID:
          type: string
          description: The unique identifier for this WireGuard panel instance.
          example: "wg-panel-001"
        wgIfPrefix:
          type: string
          description: The prefix used for WireGuard interface names.
          example: "wg"
    Interface:
      type: object
      properties:
        id:
          type: string
          description: Unique ID of the interface. Allocated on creation.
          readOnly: true
        ifname:
          type: string
          description: A short name for the interface. The actual device name will be `wg-[ifname]`. Must be unique.
          pattern: '^[A-Za-z0-9_-]{1,12}$'
          example: 'prod'
        enabled:
          type: boolean
          description: Reflects the current operational state of the interface. Use the set-enable endpoint to change.
          readOnly: true
        vrfName:
          type: string
          description: The VRF (Virtual Routing and Forwarding) to which this interface belongs. If null, not in a VRF.
          nullable: true
          example: 'vrf-red'
        fwMark:
          type: string
          description: The FwMark for the WireGuard interface. Can be decimal or hexadecimal (e.g., '0xca6c'). If null, not set.
          nullable: true
          example: "51820"
        endpoint:
          type: string
          description: The public endpoint (IP or domain name) of the WireGuard interface.
          example: "vpn.example.com"
        port:
          type: integer
          description: The listen port of the WireGuard interface.
          example: 51820
        mtu:
          type: integer
          description: MTU for the WireGuard interface.
          example: 1420
        publicKey:
          type: string
          description: The public key of the interface, derived from the private key.
          readOnly: true
        servers:
          type: array
          description: List of servers belonging to this interface. Included on a GET request for a single interface.
          items:
            $ref: '#/components/schemas/Server'
          readOnly: true
      required:
        - id
        - ifname
        - enabled
        - endpoint
        - port
        - publicKey
    InterfaceCreate:
      type: object
      properties:
        ifname:
          type: string
          description: A short name for the interface. The actual device name will be `wg-[ifname]`. Must be unique.
          pattern: '^[A-Za-z0-9_-]{1,12}$'
          example: 'prod'
        vrfName:
          type: string
          description: The VRF to which this interface belongs.
          nullable: true
        fwMark:
          type: string
          description: The FwMark for the WireGuard interface. Can be decimal or hexadecimal.
          nullable: true
        endpoint:
          type: string
          description: The public endpoint (IP or domain name) of the WireGuard interface.
          example: "vpn.example.com"
        port:
          type: integer
          description: The listen port of the WireGuard interface.
          example: 51820
        mtu:
          type: integer
          description: MTU for the WireGuard interface.
          default: 1420
        privateKey:
          type: string
          description: The private key for the interface. If null, a new key will be generated.
          nullable: true
          writeOnly: true
      required:
        - ifname
        - endpoint
        - port
    InterfaceUpdate:
      type: object
      properties:
        ifname:
          type: string
          description: A short name for the interface.
          pattern: '^[A-Za-z0-9_-]{1,12}$'
        vrfName:
          type: string
          description: The VRF to which this interface belongs.
          nullable: true
        fwMark:
          type: string
          description: The FwMark for the WireGuard interface. Can be decimal or hexadecimal.
          nullable: true
        endpoint:
          type: string
          description: The public endpoint of the WireGuard interface.
        port:
          type: integer
          description: The listen port of the WireGuard interface.
        mtu:
          type: integer
          description: MTU for the WireGuard interface.
        privateKey:
          type: string
          description: The private key for the interface. If provided, the key will be updated. Set to null to keep unchanged.
          nullable: true
          writeOnly: true
    SnatConfig:
      type: object
      properties:
        enabled:
          type: boolean
          description: Enable Source NAT for clients.
        snatIpNet:
          type: string
          nullable: true
          description: |
            IP/Network to use for SNAT, in CIDR notation. The behavior depends on the IP version and format:
            - **`null`**: use MASQUERADE.
            - **IPv4**: An `IPv4/32` uses SNAT. Any other mask length is invalid.
            - **IPv6**: An `IPv6/128` uses SNAT. A network with a mask matching the server's network prefix uses NETMAP.
            When roaming is enabled:
            - **IPv4**: Must be `0.0.0.0/32` to use a dynamic public IP from the master interface.
            - **IPv6**: `::/128` uses a dynamic public IP. An `IPv6Net` (e.g., `::a:b:c:d/112`) is used as an offset from the master interface's dynamic network.
          example: "203.0.113.5/32"
        snatExcludedNetwork:
          type: string
          description: Network range to exclude from SNAT in CIDR notation. If null, defaults to the server's own network range.
          nullable: true
          example: "192.168.0.0/16"
        roamingMasterInterface:
          type: string
          maxLength: 15
          nullable: true
          description: "Enables SNAT roaming. The service will monitor this interface for IP changes and update NAT rules automatically."
          example: "eth0"
        roamingPseudoBridge:
          type: boolean
          description: "**IPv6-only:** If true, enables pseudo-bridging for the NETMAP target network on the roaming master interface."
          default: false
    ServerNetworkConfig:
      type: object
      description: Configuration for network settings (IPv4 or IPv6) on a server.
      properties:
        enabled:
          type: boolean
          description: Whether this IP version is enabled for the server.
        network:
          type: string
          description: The server's network in CIDR notation (e.g., 10.0.0.1/24 or fd00::1/64). Defines the server's IP and the pool for clients.
          example: "10.0.0.1/24"
        pseudoBridgeMasterInterface:
          type: string
          maxLength: 15
          nullable: true
          description: If set, enables a pseudo-bridge on the specified interface to handle ARP (IPv4) or Neighbor Discovery (IPv6) requests for client IPs.
          example: "br-lan"
        snat:
          $ref: '#/components/schemas/SnatConfig'
        routedNetworks:
          type: array
          items:
            type: string
            description: A network in CIDR notation (e.g., 10.0.0.0/24).
          description: List of networks to route for clients (equivalent to AllowedIPs). If null, defaults to server's own network range. Must contain server's own network range if specified.
          example: ["10.0.0.0/24", "172.16.0.0/16"]
          nullable: true
        routedNetworksFirewall:
          type: boolean
          description: If true, add firewall rules to strictly allow traffic only to the specified routed networks (unless 0.0.0.0/0 or ::/0 is used).
    Server:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
        enabled:
          type: boolean
          readOnly: true
          description: Reflects the current operational state of the server. Use the set-enable endpoint to change. Defaults to false on creation.
        dns:
          type: array
          items:
            type: string
            format: ip
          description: "List of DNS servers configured for clients of this server. If null, DNS row is omitted from client config."
          nullable: true
          example: ["1.1.1.1", "8.8.8.8"]
        ipv4:
          $ref: '#/components/schemas/ServerNetworkConfig'
        ipv6:
          $ref: '#/components/schemas/ServerNetworkConfig'
      required:
        - id
        - name
        - enabled
    ServerCreate:
      type: object
      description: At least one of ipv4 or ipv6 must be enabled and configured.
      properties:
        name:
          type: string
          example: "Default Server"
        dns:
          type: array
          items:
            type: string
            format: ip
          description: "List of DNS servers to be used by clients. Can be null."
          nullable: true
          example: ["1.1.1.1", "1.0.0.1"]
        ipv4:
          $ref: '#/components/schemas/ServerNetworkConfig'
        ipv6:
          $ref: '#/components/schemas/ServerNetworkConfig'
      required:
        - name
    ServerUpdate:
      type: object
      description: At least one of ipv4 or ipv6 must be enabled and configured.
      properties:
        name:
          type: string
        dns:
          type: array
          items:
            type: string
            format: ip
          description: "List of DNS servers to be used by clients. Can be null."
          nullable: true
        ipv4:
          $ref: '#/components/schemas/ServerNetworkConfig'
        ipv6:
          $ref: '#/components/schemas/ServerNetworkConfig'
...